- Add Clang 8 and GCC 9 to CI - Switch clang-tidy to Clang 9 - Fix compiler and clang-tidy warnings produced by the previous points

+++ src/cryfs/filesystem/fsblobstore/utils/DirEntryList.cpp
@@ -175,7 +175,7 @@ vector<DirEntry>::iterator DirEntryList::_findFirst(const BlockId &hint, std::fu
         return _entries.end();
     }
     double startpos_percent = static_cast<double>(*static_cast<const unsigned char*>(hint.data().data())) / std::numeric_limits<unsigned char>::max();
-    auto iter = _entries.begin() + static_cast<int>(startpos_percent * (_entries.size()-1));
+    auto iter = _entries.begin() + static_cast<int>(startpos_percent * static_cast<double>(_entries.size()-1));
     ASSERT(iter >= _entries.begin() && iter < _entries.end(), "Startpos out of range");
     while(iter != _entries.begin() && pred(*iter)) {
         --iter;
